backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
google-genai
google-cloud-speech
python-multipart
